feat:create home page widget
